Feature comments

+++ main/src/main/java/ru/practicum/exploreWithMe/service/CategoryServiceImpl.java
@@ -7,9 +7,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.exploreWithMe.exception.ConflictException;
 import ru.practicum.exploreWithMe.exception.NotFoundException;
 import ru.practicum.exploreWithMe.model.Category;
 import ru.practicum.exploreWithMe.repository.CategoryRepository;
+import ru.practicum.exploreWithMe.repository.EventRepository;
 
 import java.util.List;
 
@@ -20,6 +22,7 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     @Transactional
@@ -33,6 +36,9 @@ public class CategoryServiceImpl implements CategoryService {
     public Category update(Category category) {
         Category categUpd = categoryRepository.findById(category.getId())
                 .orElseThrow(() -> new NotFoundException(HttpStatus.NOT_FOUND));
+        if (category.getName().equals(categUpd.getName())) {
+            throw new ConflictException(HttpStatus.CONFLICT);
+        }
         categUpd.setName(category.getName());
         log.info("---===>>>CATG_SERV updated category=/{}/", category);
         return categUpd;
@@ -55,6 +61,9 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional
     public void remove(long id) {
         if (categoryRepository.findById(id).isPresent()) {
+            if (eventRepository.getCountByCategory(id) > 0) {
+                throw new ConflictException(HttpStatus.CONFLICT);
+            }
             log.info("deleted category id=/{}/", id);
             categoryRepository.deleteById(id);
         } else {
